+++ modules/sys/sockets.js
@@ -11,12 +11,19 @@ export class HMSocket {
 
 export const SOCKET_TYPES = {
     DRAW_REACH: 0,
+    RENDER_ACTORS: 1,
+};
+
+export const handleSocketEvent = ({type, data, userId}) => {
+    if (type === SOCKET_TYPES.DRAW_REACH) drawReach(data, userId);
+    if (type === SOCKET_TYPES.RENDER_APPS) renderApps(data, userId);
 };
 
 function drawReach(tokenId) {
     game.canvas.tokens.get(tokenId).drawReach();
 }
 
-export const handleSocketEvent = ({type, data, userId}) => {
-    if (type === SOCKET_TYPES.DRAW_REACH) drawReach(data, userId);
-};
+function renderApps() {
+    const apps = game.actors.reduce((acc, actor) => ({...acc, ...actor.apps}), {});
+    Object.keys(apps).forEach((key) => apps[key].render());
+}